Start moving examples into separate file
and separate test package to ensure public API usage
