+++ app_test.go
@@ -7,7 +7,6 @@ import (
 	"flag"
 	"fmt"
 	"io"
-	"net/mail"
 	"os"
 	"reflect"
 	"strconv"
@@ -33,494 +32,6 @@ func init() {
 
 type opCounts struct {
 	Total, ShellComplete, OnUsageError, Before, CommandNotFound, Action, After, SubCommand int
-}
-
-func ExampleCommand_Run() {
-	// set args for examples sake
-	os.Args = []string{"greet", "--name", "Jeremy"}
-
-	cmd := &Command{
-		Name: "greet",
-		Flags: []Flag{
-			&StringFlag{Name: "name", Value: "bob", Usage: "a name to say"},
-		},
-		Action: func(c *Context) error {
-			fmt.Printf("Hello %v\n", c.String("name"))
-			return nil
-		},
-		UsageText: "app [first_arg] [second_arg]",
-		Authors:   []any{&mail.Address{Name: "Oliver Allen", Address: "oliver@toyshop.example.com"}, "gruffalo@soup-world.example.org"},
-	}
-
-	ctx, cancel := context.WithTimeout(context.Background(), 100*time.Millisecond)
-	defer cancel()
-
-	if err := cmd.Run(ctx, os.Args); err != nil {
-		return
-	}
-	// Output:
-	// Hello Jeremy
-}
-
-func ExampleCommand_Run_subcommand() {
-	// set args for examples sake
-	os.Args = []string{"say", "hi", "english", "--name", "Jeremy"}
-	cmd := &Command{
-		Name: "say",
-		Commands: []*Command{
-			{
-				Name:        "hello",
-				Aliases:     []string{"hi"},
-				Usage:       "use it to see a description",
-				Description: "This is how we describe hello the function",
-				Commands: []*Command{
-					{
-						Name:        "english",
-						Aliases:     []string{"en"},
-						Usage:       "sends a greeting in english",
-						Description: "greets someone in english",
-						Flags: []Flag{
-							&StringFlag{
-								Name:  "name",
-								Value: "Bob",
-								Usage: "Name of the person to greet",
-							},
-						},
-						Action: func(c *Context) error {
-							fmt.Println("Hello,", c.String("name"))
-							return nil
-						},
-					},
-				},
-			},
-		},
-	}
-
-	ctx, cancel := context.WithTimeout(context.Background(), 100*time.Millisecond)
-	defer cancel()
-
-	_ = cmd.Run(ctx, os.Args)
-	// Output:
-	// Hello, Jeremy
-}
-
-func ExampleCommand_Run_appHelp() {
-	cmd := &Command{
-		Name:        "greet",
-		Version:     "0.1.0",
-		Description: "This is how we describe greet the app",
-		Authors: []any{
-			&mail.Address{Name: "Harrison", Address: "harrison@lolwut.example.com"},
-			"Oliver Allen  <oliver@toyshop.example.com>",
-		},
-		Flags: []Flag{
-			&StringFlag{Name: "name", Value: "bob", Usage: "a name to say"},
-		},
-		Commands: []*Command{
-			{
-				Name:        "describeit",
-				Aliases:     []string{"d"},
-				Usage:       "use it to see a description",
-				Description: "This is how we describe describeit the function",
-				Action: func(*Context) error {
-					fmt.Printf("i like to describe things")
-					return nil
-				},
-			},
-		},
-	}
-
-	ctx, cancel := context.WithTimeout(context.Background(), 100*time.Millisecond)
-	defer cancel()
-
-	// Simulate the command line arguments
-	os.Args = []string{"greet", "help"}
-
-	_ = cmd.Run(ctx, os.Args)
-	// Output:
-	// NAME:
-	//    greet - A new cli application
-	//
-	// USAGE:
-	//    greet [global options] [command [command options]] [arguments...]
-	//
-	// VERSION:
-	//    0.1.0
-	//
-	// DESCRIPTION:
-	//    This is how we describe greet the app
-	//
-	// AUTHORS:
-	//    "Harrison" <harrison@lolwut.example.com>
-	//    Oliver Allen  <oliver@toyshop.example.com>
-	//
-	// COMMANDS:
-	//    describeit, d  use it to see a description
-	//    help, h        Shows a list of commands or help for one command
-	//
-	// GLOBAL OPTIONS:
-	//    --name value   a name to say (default: "bob")
-	//    --help, -h     show help (default: false)
-	//    --version, -v  print the version (default: false)
-}
-
-func ExampleCommand_Run_commandHelp() {
-	cmd := &Command{
-		Name: "greet",
-		Flags: []Flag{
-			&StringFlag{Name: "name", Value: "bob", Usage: "a name to say"},
-		},
-		Action: func(cCtx *Context) error {
-			fmt.Fprintf(cCtx.Command.Root().Writer, "hello to %[1]q\n", cCtx.String("name"))
-			return nil
-		},
-		Commands: []*Command{
-			{
-				Name:        "describeit",
-				Aliases:     []string{"d"},
-				Usage:       "use it to see a description",
-				Description: "This is how we describe describeit the function",
-				Action: func(*Context) error {
-					fmt.Println("i like to describe things")
-					return nil
-				},
-			},
-		},
-	}
-
-	ctx, cancel := context.WithTimeout(context.Background(), 100*time.Millisecond)
-	defer cancel()
-
-	// Simulate the command line arguments
-	os.Args = []string{"greet", "h", "describeit"}
-
-	_ = cmd.Run(ctx, os.Args)
-	// Output:
-	// NAME:
-	//    greet describeit - use it to see a description
-	//
-	// USAGE:
-	//    greet describeit [command [command options]] [arguments...]
-	//
-	// DESCRIPTION:
-	//    This is how we describe describeit the function
-	//
-	// COMMANDS:
-	//    help, h  Shows a list of commands or help for one command
-	//
-	// OPTIONS:
-	//    --help, -h  show help (default: false)
-}
-
-func ExampleCommand_Run_noAction() {
-	cmd := &Command{
-		Name: "greet",
-	}
-
-	ctx, cancel := context.WithTimeout(context.Background(), 100*time.Millisecond)
-	defer cancel()
-
-	_ = cmd.Run(ctx, []string{"greet"})
-	// Output:
-	// NAME:
-	//    greet - A new cli application
-	//
-	// USAGE:
-	//    greet [global options] [command [command options]] [arguments...]
-	//
-	// COMMANDS:
-	//    help, h  Shows a list of commands or help for one command
-	//
-	// GLOBAL OPTIONS:
-	//    --help, -h  show help (default: false)
-}
-
-func ExampleCommand_Run_subcommandNoAction() {
-	cmd := &Command{
-		Name: "greet",
-		Commands: []*Command{
-			{
-				Name:        "describeit",
-				Aliases:     []string{"d"},
-				Usage:       "use it to see a description",
-				Description: "This is how we describe describeit the function",
-			},
-		},
-	}
-
-	ctx, cancel := context.WithTimeout(context.Background(), 100*time.Millisecond)
-	defer cancel()
-
-	_ = cmd.Run(ctx, []string{"greet", "describeit"})
-	// Output:
-	// NAME:
-	//    greet describeit - use it to see a description
-	//
-	// USAGE:
-	//    greet describeit [command [command options]] [arguments...]
-	//
-	// DESCRIPTION:
-	//    This is how we describe describeit the function
-	//
-	// OPTIONS:
-	//    --help, -h  show help (default: false)
-}
-
-func ExampleCommand_Run_bashComplete_withShortFlag() {
-	os.Setenv("SHELL", "bash")
-	os.Args = []string{"greet", "-", "--generate-shell-completion"}
-
-	cmd := &Command{
-		Name:                  "greet",
-		EnableShellCompletion: true,
-		Flags: []Flag{
-			&IntFlag{
-				Name:    "other",
-				Aliases: []string{"o"},
-			},
-			&StringFlag{
-				Name:    "xyz",
-				Aliases: []string{"x"},
-			},
-		},
-	}
-
-	ctx, cancel := context.WithTimeout(context.Background(), 100*time.Millisecond)
-	defer cancel()
-
-	_ = cmd.Run(ctx, os.Args)
-	// Output:
-	// --other
-	// -o
-	// --xyz
-	// -x
-	// --help
-	// -h
-}
-
-func ExampleCommand_Run_bashComplete_withLongFlag() {
-	os.Setenv("SHELL", "bash")
-	os.Args = []string{"greet", "--s", "--generate-shell-completion"}
-
-	cmd := &Command{
-		Name:                  "greet",
-		EnableShellCompletion: true,
-		Flags: []Flag{
-			&IntFlag{
-				Name:    "other",
-				Aliases: []string{"o"},
-			},
-			&StringFlag{
-				Name:    "xyz",
-				Aliases: []string{"x"},
-			},
-			&StringFlag{
-				Name: "some-flag,s",
-			},
-			&StringFlag{
-				Name: "similar-flag",
-			},
-		},
-	}
-
-	ctx, cancel := context.WithTimeout(context.Background(), 100*time.Millisecond)
-	defer cancel()
-
-	_ = cmd.Run(ctx, os.Args)
-	// Output:
-	// --some-flag
-	// --similar-flag
-}
-
-func ExampleCommand_Run_bashComplete_withMultipleLongFlag() {
-	os.Setenv("SHELL", "bash")
-	os.Args = []string{"greet", "--st", "--generate-shell-completion"}
-
-	cmd := &Command{
-		Name:                  "greet",
-		EnableShellCompletion: true,
-		Flags: []Flag{
-			&IntFlag{
-				Name:    "int-flag",
-				Aliases: []string{"i"},
-			},
-			&StringFlag{
-				Name:    "string",
-				Aliases: []string{"s"},
-			},
-			&StringFlag{
-				Name: "string-flag-2",
-			},
-			&StringFlag{
-				Name: "similar-flag",
-			},
-			&StringFlag{
-				Name: "some-flag",
-			},
-		},
-	}
-
-	ctx, cancel := context.WithTimeout(context.Background(), 100*time.Millisecond)
-	defer cancel()
-
-	_ = cmd.Run(ctx, os.Args)
-	// Output:
-	// --string
-	// --string-flag-2
-}
-
-func ExampleCommand_Run_bashComplete() {
-	os.Setenv("SHELL", "bash")
-	os.Args = []string{"greet", "--generate-shell-completion"}
-
-	cmd := &Command{
-		Name:                  "greet",
-		EnableShellCompletion: true,
-		Commands: []*Command{
-			{
-				Name:        "describeit",
-				Aliases:     []string{"d"},
-				Usage:       "use it to see a description",
-				Description: "This is how we describe describeit the function",
-				Action: func(*Context) error {
-					fmt.Printf("i like to describe things")
-					return nil
-				},
-			}, {
-				Name:        "next",
-				Usage:       "next example",
-				Description: "more stuff to see when generating shell completion",
-				Action: func(*Context) error {
-					fmt.Printf("the next example")
-					return nil
-				},
-			},
-		},
-	}
-
-	ctx, cancel := context.WithTimeout(context.Background(), 100*time.Millisecond)
-	defer cancel()
-
-	_ = cmd.Run(ctx, os.Args)
-	// Output:
-	// describeit
-	// d
-	// next
-	// help
-	// h
-}
-
-func ExampleCommand_Run_zshComplete() {
-	// set args for examples sake
-	os.Args = []string{"greet", "--generate-shell-completion"}
-	_ = os.Setenv("SHELL", "/usr/bin/zsh")
-
-	cmd := &Command{
-		Name:                  "greet",
-		EnableShellCompletion: true,
-		Commands: []*Command{
-			{
-				Name:        "describeit",
-				Aliases:     []string{"d"},
-				Usage:       "use it to see a description",
-				Description: "This is how we describe describeit the function",
-				Action: func(*Context) error {
-					fmt.Printf("i like to describe things")
-					return nil
-				},
-			}, {
-				Name:        "next",
-				Usage:       "next example",
-				Description: "more stuff to see when generating bash completion",
-				Action: func(*Context) error {
-					fmt.Printf("the next example")
-					return nil
-				},
-			},
-		},
-	}
-
-	ctx, cancel := context.WithTimeout(context.Background(), 100*time.Millisecond)
-	defer cancel()
-
-	_ = cmd.Run(ctx, os.Args)
-	// Output:
-	// describeit:use it to see a description
-	// d:use it to see a description
-	// next:next example
-	// help:Shows a list of commands or help for one command
-	// h:Shows a list of commands or help for one command
-}
-
-func ExampleCommand_Run_sliceValues() {
-	// set args for examples sake
-	os.Args = []string{
-		"multi_values",
-		"--stringSclice", "parsed1,parsed2", "--stringSclice", "parsed3,parsed4",
-		"--float64Sclice", "13.3,14.4", "--float64Sclice", "15.5,16.6",
-		"--int64Sclice", "13,14", "--int64Sclice", "15,16",
-		"--intSclice", "13,14", "--intSclice", "15,16",
-	}
-	cmd := &Command{
-		Name: "multi_values",
-		Flags: []Flag{
-			&StringSliceFlag{Name: "stringSclice"},
-			&Float64SliceFlag{Name: "float64Sclice"},
-			&Int64SliceFlag{Name: "int64Sclice"},
-			&IntSliceFlag{Name: "intSclice"},
-		},
-	}
-	cmd.Action = func(ctx *Context) error {
-		for i, v := range ctx.FlagNames() {
-			fmt.Printf("%d-%s %#v\n", i, v, ctx.Value(v))
-		}
-		err := ctx.Err()
-		fmt.Println("error:", err)
-		return err
-	}
-
-	ctx, cancel := context.WithTimeout(context.Background(), 100*time.Millisecond)
-	defer cancel()
-
-	_ = cmd.Run(ctx, os.Args)
-	// Output:
-	// 0-float64Sclice []float64{13.3, 14.4, 15.5, 16.6}
-	// 1-int64Sclice []int64{13, 14, 15, 16}
-	// 2-intSclice []int{13, 14, 15, 16}
-	// 3-stringSclice []string{"parsed1", "parsed2", "parsed3", "parsed4"}
-	// error: <nil>
-}
-
-func ExampleCommand_Run_mapValues() {
-	// set args for examples sake
-	os.Args = []string{
-		"multi_values",
-		"--stringMap", "parsed1=parsed two", "--stringMap", "parsed3=",
-	}
-	cmd := &Command{
-		Name: "multi_values",
-		Flags: []Flag{
-			&StringMapFlag{Name: "stringMap"},
-		},
-		Action: func(ctx *Context) error {
-			for i, v := range ctx.FlagNames() {
-				fmt.Printf("%d-%s %#v\n", i, v, ctx.StringMap(v))
-			}
-			fmt.Printf("notfound %#v\n", ctx.StringMap("notfound"))
-			err := ctx.Err()
-			fmt.Println("error:", err)
-			return err
-		},
-	}
-
-	ctx, cancel := context.WithTimeout(context.Background(), 100*time.Millisecond)
-	defer cancel()
-
-	_ = cmd.Run(ctx, os.Args)
-	// Output:
-	// 0-stringMap map[string]string{"parsed1":"parsed two", "parsed3":""}
-	// notfound map[string]string(nil)
-	// error: <nil>
 }
 
 func TestApp_Run(t *testing.T) {
